Split homeassistant_alerts constants and coordinator
