+++ homeassistant/components/homeassistant_alerts/const.py
@@ -1,0 +1,13 @@
+"""Constants for the Home Assistant alerts integration."""
+
+from __future__ import annotations
+
+from datetime import timedelta
+
+import aiohttp
+
+COMPONENT_LOADED_COOLDOWN = 30
+DOMAIN = "homeassistant_alerts"
+UPDATE_INTERVAL = timedelta(hours=3)
+
+REQUEST_TIMEOUT = aiohttp.ClientTimeout(total=30)
